# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Bemantis API",
    "description": "Plataforma de gestion de leads con integracion WhatsApp IA",
    "version": "1.0.0"
  },
  "paths": {
...
    "/quotes": {
      "get": {
        "tags": [
          "Cotizaciones"
        ],
        "summary": "List Quotes",
        "operationId": "list_quotes_quotes_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/QuoteStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "contact_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Contact Id"
            }
          },
          {
            "name": "seller_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Seller Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/QuoteResponse"
                  },
                  "title": "Response List Quotes Quotes Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Cotizaciones"
        ],
        "summary": "Create Quote",
        "operationId": "create_quote_quotes_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuoteCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuoteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/quotes/{quote_id}": {
      "get": {
        "tags": [
          "Cotizaciones"
        ],
        "summary": "Get Quote",
        "operationId": "get_quote_quotes__quote_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "quote_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Quote Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuoteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Cotizaciones"
        ],
        "summary": "Update Quote",
        "operationId": "update_quote_quotes__quote_id__put",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "quote_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Quote Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuoteUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuoteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/quotes/{quote_id}/status": {
      "patch": {
        "tags": [
          "Cotizaciones"
        ],
        "summary": "Update Quote Status",
        "operationId": "update_quote_status_quotes__quote_id__status_patch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "quote_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Quote Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuoteStatusUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuoteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/quotes/{quote_id}/items": {
      "post": {
        "tags": [
          "Cotizaciones"
        ],
        "summary": "Add Quote Item",
        "operationId": "add_quote_item_quotes__quote_id__items_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "quote_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Quote Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuoteItemCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuoteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/quotes/{quote_id}/items/{item_id}": {
      "put": {
        "tags": [
          "Cotizaciones"
        ],
        "summary": "Update Quote Item",
        "operationId": "update_quote_item_quotes__quote_id__items__item_id__put",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "quote_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Quote Id"
            }
          },
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Item Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuoteItemUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuoteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Cotizaciones"
        ],
        "summary": "Delete Quote Item",
        "operationId": "delete_quote_item_quotes__quote_id__items__item_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "quote_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Quote Id"
            }
          },
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Item Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuoteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ContactBrief": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "company": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "email",
          "company"
        ],
        "title": "ContactBrief"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "QuoteCreate": {
        "properties": {
          "contact_id": {
            "type": "integer",
            "title": "Contact Id"
          },
          "discount_percent": {
            "type": "number",
            "title": "Discount Percent",
            "default": 0
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "items": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/QuoteItemCreate"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "contact_id"
        ],
        "title": "QuoteCreate"
      },
      "QuoteItemCreate": {
        "properties": {
          "service_id": {
            "type": "integer",
            "title": "Service Id"
          },
          "custom_price": {
            "type": "number",
            "title": "Custom Price"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "service_id",
          "custom_price"
        ],
        "title": "QuoteItemCreate"
      },
      "QuoteItemResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "quote_id": {
            "type": "integer",
            "title": "Quote Id"
          },
          "service_id": {
            "type": "integer",
            "title": "Service Id"
          },
          "custom_price": {
            "type": "number",
            "title": "Custom Price"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "service": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ServiceBrief"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "quote_id",
          "service_id",
          "custom_price",
          "notes",
          "service"
        ],
        "title": "QuoteItemResponse"
      },
      "QuoteItemUpdate": {
        "properties": {
          "custom_price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Custom Price"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "title": "QuoteItemUpdate"
      },
      "QuoteResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "quote_number": {
            "type": "string",
            "title": "Quote Number"
          },
          "contact_id": {
            "type": "integer",
            "title": "Contact Id"
          },
          "seller_id": {
            "type": "integer",
            "title": "Seller Id"
          },
          "status": {
            "$ref": "#/components/schemas/QuoteStatus"
          },
          "discount_percent": {
            "type": "number",
            "title": "Discount Percent"
          },
          "tax_rate": {
            "type": "number",
            "title": "Tax Rate"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "subtotal": {
            "type": "number",
            "title": "Subtotal"
          },
          "discount_amount": {
            "type": "number",
            "title": "Discount Amount"
          },
          "tax_amount": {
            "type": "number",
            "title": "Tax Amount"
          },
          "total": {
            "type": "number",
            "title": "Total"
          },
          "contact": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ContactBrief"
              },
              {
                "type": "null"
              }
            ]
          },
          "seller": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SellerBrief"
              },
              {
                "type": "null"
              }
            ]
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/QuoteItemResponse"
            },
            "type": "array",
            "title": "Items",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "quote_number",
          "contact_id",
          "seller_id",
          "status",
          "discount_percent",
          "tax_rate",
          "notes",
          "created_at",
          "subtotal",
          "discount_amount",
          "tax_amount",
          "total",
          "contact",
          "seller"
        ],
        "title": "QuoteResponse"
      },
      "QuoteStatus": {
        "type": "string",
        "enum": [
          "borrador",
          "enviada",
          "aprobada",
          "rechazada"
        ],
        "title": "QuoteStatus"
      },
      "QuoteStatusUpdate": {
        "properties": {
          "status": {
            "$ref": "#/components/schemas/QuoteStatus"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "QuoteStatusUpdate"
      },
      "QuoteUpdate": {
        "properties": {
          "discount_percent": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Discount Percent"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "title": "QuoteUpdate"
      },
...
      "SellerBrief": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "nombre": {
            "type": "string",
            "title": "Nombre"
          },
          "correo": {
            "type": "string",
            "title": "Correo"
          }
        },
        "type": "object",
        "required": [
          "id",
          "nombre",
          "correo"
        ],
        "title": "SellerBrief"
      },
      "ServiceBrief": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "category": {
            "type": "string",
            "title": "Category"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "category"
        ],
        "title": "ServiceBrief"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {

            },
            "tokenUrl": "/auth/login"
          }
        }
      }
    }
  }
}